Removed unecessary example duplicates (i.e. x and x-views)
