--- linera-examples/crowd-funding-views/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "crowd-funding-views"
-version = "0.1.0"
-authors = ["Linera <contact@linera.io>"]
-edition = "2021"
-
-[dependencies]
-async-trait = "0.1.52"
-bcs = "0.1.3"
-crowd-funding = { path = "../crowd-funding" }
-fungible = { path = "../fungible" }
-linera-sdk = { path = "../../linera-sdk" }
-linera-views = { path = "../../linera-views" }
-serde = { version = "1.0.130", features = ["derive"] }
-thiserror = "1.0.31"
-
-[[bin]]
-name = "crowd_funding_views_contract"
-path = "src/contract.rs"
-
-[[bin]]
-name = "crowd_funding_views_service"
-path = "src/service.rs"
