+++ linera-examples/crowd-funding/Cargo.toml
@@ -9,13 +9,14 @@ async-trait = "0.1.52"
 bcs = "0.1.3"
 fungible = { path = "../fungible" }
 linera-sdk = { path = "../../linera-sdk" }
-serde = { version = "1.0.130", features = ["derive"] }
+linera-views = { path = "../../linera-views" }
+serde = {version = "1.0.160", features = ["derive"] }
 thiserror = "1.0.31"
 
 [[bin]]
-name = "crowd-funding_contract"
+name = "crowd_funding_contract"
 path = "src/contract.rs"
 
 [[bin]]
-name = "crowd-funding_service"
+name = "crowd_funding_service"
 path = "src/service.rs"
