--- linera-examples/fungible-views/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "fungible-views"
-version = "0.1.0"
-authors = ["Linera <contact@linera.io>"]
-edition = "2021"
-
-[dependencies]
-async-graphql = { version = "5.0.6", default-features = false }
-async-trait = "0.1.52"
-bcs = "0.1.3"
-fungible = { path = "../fungible" }
-linera-sdk = { path = "../../linera-sdk" }
-linera-views = { path = "../../linera-views" }
-thiserror = "1.0.31"
-serde_json = "1.0.93"
-
-[[bin]]
-name = "fungible_views_contract"
-path = "src/contract.rs"
-
-[[bin]]
-name = "fungible_views_service"
-path = "src/service.rs"
