+++ linera-examples/fungible/Cargo.toml
@@ -9,12 +9,10 @@ async-graphql = { version = "5.0.6", default-features = false }
 async-trait = "0.1.52"
 bcs = "0.1.3"
 linera-sdk = { path = "../../linera-sdk" }
-serde = { version = "1.0.130", features = ["derive"] }
+linera-views = { path = "../../linera-views" }
 thiserror = "1.0.31"
-
-[target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
-linera-sdk = { path = "../../linera-sdk", features = ["test", "wasmer", "wasmtime"] }
-tokio = { version = "1.25.0", features = ["macros", "rt-multi-thread"] }
+serde_json = "1.0.93"
+serde = { version = "1.0.160", features = ["derive"] }
 
 [[bin]]
 name = "fungible_contract"
