--- linera-examples/fungible2/web-frontend/src/GraphQLProvider.js
@@ -1,47 +1,0 @@
-import {
-  ApolloClient,
-  ApolloProvider,
-  HttpLink,
-  InMemoryCache,
-  split,
-} from "@apollo/client";
-import React from "react";
-import { GraphQLWsLink } from "@apollo/client/link/subscriptions";
-import { createClient } from "graphql-ws";
-import { getMainDefinition } from "@apollo/client/utilities";
-
-function GraphQLProvider({ applicationId, port, children }) {
-  let client = apolloClient(applicationId, port);
-  return <ApolloProvider client={client}>{children}</ApolloProvider>;
-}
-
-function apolloClient(applicationId, port) {
-  const wsLink = new GraphQLWsLink(
-    createClient({
-      url: `ws://localhost:${port}/ws`,
-    })
-  );
-
-  const httpLink = new HttpLink({
-    uri: `http://localhost:${port}/applications/` + applicationId,
-  });
-
-  const splitLink = split(
-    ({ query }) => {
-      const definition = getMainDefinition(query);
-      return (
-        definition.kind === "OperationDefinition" &&
-        definition.operation === "subscription"
-      );
-    },
-    wsLink,
-    httpLink
-  );
-
-  return new ApolloClient({
-    link: splitLink,
-    cache: new InMemoryCache(),
-  });
-}
-
-export default GraphQLProvider;
