--- linera-examples/reentrant-counter-views/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "reentrant-counter-views"
-version = "0.1.0"
-authors = ["Linera <contact@linera.io>"]
-edition = "2021"
-
-[dependencies]
-async-trait = "0.1.52"
-bcs = "0.1.3"
-futures = "0.3.17"
-linera-sdk = { path = "../../linera-sdk" }
-linera-views = { path = "../../linera-views" }
-thiserror = "1.0.31"
-
-[dev-dependencies]
-linera-sdk = { path = "../../linera-sdk", features = ["test"] }
-webassembly-test = "0.1.0"
-
-[[bin]]
-name = "reentrant_counter_views_contract"
-path = "src/contract.rs"
-
-[[bin]]
-name = "reentrant_counter_views_service"
-path = "src/service.rs"
