+++ linera-examples/reentrant-counter/Cargo.toml
@@ -9,7 +9,7 @@ async-trait = "0.1.52"
 bcs = "0.1.3"
 futures = "0.3.17"
 linera-sdk = { path = "../../linera-sdk" }
-serde = { version = "1.0.130", features = ["derive"] }
+linera-views = { path = "../../linera-views" }
 thiserror = "1.0.31"
 
 [dev-dependencies]
