+++ linera-examples/reentrant-counter/src/service.rs
@@ -7,24 +7,32 @@ mod state;
 
 use self::state::ReentrantCounter;
 use async_trait::async_trait;
-use linera_sdk::{QueryContext, Service, SimpleStateStorage};
+use linera_sdk::{
+    service::system_api::ReadOnlyViewStorageContext, QueryContext, Service, ViewStateStorage,
+};
+use linera_views::{common::Context, views::ViewError};
 use std::sync::Arc;
 use thiserror::Error;
 
-linera_sdk::service!(ReentrantCounter);
+linera_sdk::service!(ReentrantCounter<ReadOnlyViewStorageContext>);
 
 #[async_trait]
-impl Service for ReentrantCounter {
+impl<C> Service for ReentrantCounter<C>
+where
+    C: Context + Send + Sync + Clone + 'static,
+    ViewError: From<C::Error>,
+{
     type Error = Error;
-    type Storage = SimpleStateStorage<Self>;
+    type Storage = ViewStateStorage<Self>;
 
     async fn query_application(
         self: Arc<Self>,
         _context: &QueryContext,
         argument: &[u8],
     ) -> Result<Vec<u8>, Self::Error> {
+        let value = self.value.get();
         match argument {
-            &[] => Ok(bcs::to_bytes(&self.value).expect("Serialization should not fail")),
+            &[] => Ok(bcs::to_bytes(&value).expect("Serialization should not fail")),
             _ => Err(Error::InvalidQuery),
         }
     }
@@ -40,17 +48,24 @@ pub enum Error {
 
 #[cfg(test)]
 mod tests {
-    use super::{Error, ReentrantCounter};
+    use super::Error;
+    use crate::ReentrantCounter;
     use futures::FutureExt;
     use linera_sdk::{base::ChainId, QueryContext, Service};
+    use linera_views::{memory::create_test_context, views::View};
     use std::sync::Arc;
     use webassembly_test::webassembly_test;
 
     #[webassembly_test]
     fn query() {
         let value = 61_098_721_u128;
-        let counter = Arc::new(ReentrantCounter { value });
-
+        let context = create_test_context();
+        let mut counter = ReentrantCounter::load(context)
+            .now_or_never()
+            .unwrap()
+            .expect("Failed to load Counter");
+        counter.value.set(value);
+        let counter = Arc::new(counter);
         let result = counter
             .query_application(&dummy_query_context(), &[])
             .now_or_never()
@@ -65,7 +80,13 @@ mod tests {
     #[webassembly_test]
     fn invalid_query() {
         let value = 4_u128;
-        let counter = Arc::new(ReentrantCounter { value });
+        let context = create_test_context();
+        let mut counter = ReentrantCounter::load(context)
+            .now_or_never()
+            .unwrap()
+            .expect("Failed to load Counter");
+        counter.value.set(value);
+        let counter = Arc::new(counter);
 
         let dummy_argument = [2];
         let result = counter
