+++ linera-examples/reentrant-counter/src/state.rs
@@ -1,10 +1,14 @@
 // Copyright (c) Zefchain Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use serde::{Deserialize, Serialize};
+use linera_views::{
+    common::Context,
+    register_view::RegisterView,
+    views::{RootView, View},
+};
 
 /// The application state.
-#[derive(Clone, Copy, Debug, Default, Deserialize, Serialize)]
-pub struct ReentrantCounter {
-    pub value: u128,
+#[derive(Debug, RootView)]
+pub struct ReentrantCounter<C> {
+    pub value: RegisterView<C, u128>,
 }
